More flexible train_pytorch.py for openpi (#4)
* More flexible train_pytorch.py for openpi

1. Allow more general batch format. Previously, a batch from
the data loader is always assumed to be a tuple of observation and action.
Since we will train with more information, the handling of a batch should
be more general.

2. Support more flexible result (i.e. dict) from model() so that more information
can be logged.

3. Support dataset wrapper and data loader wrapper so that we
can have more customization for data loading.

* Address comments

* Address comments

Also moved RemoveStrings from compute_norm_stats to transforms
as it causes issues when running compute_nomr_stats.main() from
another scripts.